[DF-12055] Fixed review issues
Former-commit-id: bf2db2c1a9fed8c996b8ae1d9a9b2806a90d017e

+++ Sources/Internal/Engine/Private/Qt/WindowBackendQt.cpp
@@ -297,7 +297,7 @@ void WindowBackend::OnCreated()
 
     WindowBackendDetails::Kostil_ForceUpdateCurrentScreen(renderWidget, engineBackend->GetNativeService()->GetApplication());
     float32 dpi = renderWidget->logicalDpiX();
-    float32 scale = renderWidget->devicePixelRatio();
+    float32 scale = static_cast<float32>(renderWidget->devicePixelRatio());
     float32 w = static_cast<float32>(renderWidget->width());
     float32 h = static_cast<float32>(renderWidget->height());
     mainDispatcher->PostEvent(MainDispatcherEvent::CreateWindowCreatedEvent(window, w, h, w * scale, h * scale, dpi));
@@ -335,7 +335,7 @@ void WindowBackend::OnFrame()
 
 void WindowBackend::OnResized(uint32 width, uint32 height)
 {
-    float32 scale = renderWidget->devicePixelRatio();
+    float32 scale = static_cast<float32>(renderWidget->devicePixelRatio());
     float32 w = static_cast<float32>(width);
     float32 h = static_cast<float32>(height);
     mainDispatcher->PostEvent(MainDispatcherEvent::CreateWindowSizeChangedEvent(window, w, h, w * scale, h * scale));
